Added validation to controller

+++ Gruvo/Gruvo/Controllers/ProfileController.cs
@@ -197,10 +197,16 @@ namespace Gruvo.Controllers
                 long userid = _tokenUserPairs.Pairs[Request.Cookies["Gruvo"]].Id;
                 long tweetId = Convert.ToInt64(Request.Headers["tweetId"]);
 
+                if(_repository.TweetDAO.CheckIfUserHasTweet(tweetId, userid))
+                {
+                    _repository.TweetDAO.DeletePost(tweetId);
 
-                _repository.TweetDAO.DeletePost(tweetId);
-
-                return Ok("Success!");
+                    return Ok("Success!");
+                }
+                else
+                {
+                    return BadRequest("Invalid tweet id!");
+                }
             }
             catch (Exception e)
             {
